requirements.txt
Pillow
keyboard
qrcode